Check on "data.repository" and "data.repository.name" added
The code took for granted that the JSON data received contained
a "repository" property. That is always the case when the request
originates from GitHub but that is not robust enough to resist the
World "Wild" Web.

The server crashed if the request did not contain that property
because the code attempted to access "data.repository.name". The
server now returns a 400 status code in that case.